Revert "Feature/logic profile screen"
