+++ app/build.gradle
@@ -46,7 +46,6 @@ dependencies {
     implementation 'androidx.navigation:navigation-ui-ktx:2.3.1'
     implementation 'com.google.android.material:material:1.2.1'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.7'
-    implementation 'com.github.bumptech.glide:glide:4.11.0'
 
     // Test
     testImplementation 'junit:junit:4.13'
@@ -58,12 +57,10 @@ dependencies {
     //Firebase
     implementation 'com.google.firebase:firebase-core:18.0.0'
     implementation 'com.google.firebase:firebase-messaging:21.0.0'
-    implementation 'com.google.android.gms:play-services-auth:18.1.0'
     implementation 'com.google.firebase:firebase-auth'
-    implementation 'com.google.android.gms:play-services-base:17.5.0'
+    implementation 'com.google.android.gms:play-services-auth:18.1.0'
+    implementation platform('com.google.firebase:firebase-bom:25.12.0')
     implementation 'com.facebook.android:facebook-login:5.15.3'
-    implementation 'com.google.firebase:firebase-storage:19.2.0'
-    implementation 'com.google.firebase:firebase-firestore:22.0.0'
 
     // Koin
     implementation "org.koin:koin-android:$koin_version"
