+++ app/src/main/AndroidManifest.xml
@@ -4,12 +4,6 @@
     package="com.albrivas.broadcastbottom">
 
     <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission
-        android:name="android.permission.WRITE_EXTERNAL_STORAGE"
-        android:maxSdkVersion="28" />
-    <uses-permission
-        android:name="android.permission.READ_EXTERNAL_STORAGE"
-        android:maxSdkVersion="28" />
 
     <application
         android:name=".ApplicationInit"
@@ -32,7 +26,7 @@
         <activity android:name=".ui.base.MainActivity" />
 
         <receiver
-            android:name="com.albrivas.broadcastbottom.domain.receiver.PushReceiver"
+            android:name=".receiver.PushReceiver"
             android:enabled="true"
             android:exported="true"
             android:permission="TODO"
@@ -43,7 +37,7 @@
         </receiver>
 
         <service
-            android:name="com.albrivas.broadcastbottom.data.service.PushMessagingService"
+            android:name=".service.PushMessagingService"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
