--- app/src/main/java/com/albrivas/broadcastbottom/common/Either.kt
@@ -1,9 +1,0 @@
-package com.albrivas.broadcastbottom.common
-
-sealed class Either<out L, out R> {
-
-    data class Left<out L>(val l: L) : Either<L, Nothing>()
-
-    data class Right<out R>(val r: R) : Either<Nothing, R>()
-
-}
