+++ app/src/main/java/com/albrivas/broadcastbottom/common/Extensions.kt
@@ -1,12 +1,7 @@
 package com.albrivas.broadcastbottom.common
 
 import android.content.Context
-import android.net.Uri
 import android.widget.Toast
-import androidx.appcompat.widget.AppCompatImageView
-import com.albrivas.broadcastbottom.R
-import com.bumptech.glide.Glide
-import com.bumptech.glide.request.RequestOptions
 
 fun Context.toast(message: String) {
     Toast.makeText(
@@ -14,16 +9,4 @@ fun Context.toast(message: String) {
         message,
         Toast.LENGTH_SHORT
     ).show()
-}
-
-fun AppCompatImageView.loadLocalUrl(drawable: Int) {
-    Glide.with(context).load(drawable)
-        .apply(RequestOptions().circleCrop())
-        .into(this)
-}
-
-fun AppCompatImageView.loadUrl(uri: Uri) {
-    Glide.with(context).load(uri)
-        .apply(RequestOptions().circleCrop().placeholder(R.drawable.ic_person))
-        .into(this)
 }
