+++ app/src/main/java/com/albrivas/broadcastbottom/common/base/BaseFragment.kt
@@ -1,11 +1,9 @@
 package com.albrivas.broadcastbottom.common.base
 
 import android.content.Intent
-import android.view.View
 import androidx.fragment.app.Fragment
 import com.albrivas.broadcastbottom.R
 import com.albrivas.broadcastbottom.ui.base.MainActivity
-import com.google.android.material.snackbar.Snackbar
 
 open class BaseFragment: Fragment() {
 
@@ -16,13 +14,5 @@ open class BaseFragment: Fragment() {
             it.overridePendingTransition(R.anim.slide_out_bottom, R.anim.slide_in_bottom)
             it.finish()
         }
-    }
-
-    fun showSnackBar(view: View, message: String) {
-        Snackbar.make(
-            view,
-            message,
-            Snackbar.LENGTH_SHORT
-        ).show()
     }
 }
