--- app/src/main/java/com/albrivas/broadcastbottom/data/repository/UserDataRepository.kt
@@ -1,40 +1,0 @@
-package com.albrivas.broadcastbottom.data.repository
-
-import android.net.Uri
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.storage.StorageReference
-
-class UserDataRepository(
-    private val databaseReference: StorageReference,
-    private val fireStoreDatabase: FirebaseFirestore,
-) {
-
-    fun uploadImage(uri: Uri, uid: String, result: (Exception?, Uri?) -> Unit) {
-        val imageProfileRef = databaseReference.child("users/$uid/profile")
-
-        imageProfileRef.putFile(uri).addOnSuccessListener {
-            it.metadata?.reference?.downloadUrl?.addOnSuccessListener { uri ->
-                result(null, uri)
-            }
-        }.addOnFailureListener {
-            result(it, null)
-        }
-    }
-
-    fun downloadImage(uid: String, result: (Exception?, Uri?) -> Unit) {
-        val imageProfileRef = databaseReference.child("users/$uid/profile")
-
-        imageProfileRef.downloadUrl.addOnSuccessListener { uri ->
-            result(null, uri)
-        }.addOnFailureListener {
-            result(it, null)
-        }
-    }
-
-    fun updateUserInformation(uid: String, user: HashMap<String, out Any?>, result: (Exception) -> Unit) {
-        fireStoreDatabase.collection("users").document(uid).set(user)
-            .addOnFailureListener {
-                result(it)
-            }
-    }
-}
