--- app/src/main/java/com/albrivas/broadcastbottom/injection/modules/DataModule.kt
@@ -1,8 +1,0 @@
-package com.albrivas.broadcastbottom.injection.modules
-
-import com.albrivas.broadcastbottom.data.repository.UserDataRepository
-import org.koin.dsl.module
-
-val dataModule = module {
-    factory { UserDataRepository(get(), get()) }
-}
