+++ app/src/main/java/com/albrivas/broadcastbottom/ui/settings/SettingsFragment.kt
@@ -1,0 +1,25 @@
+package com.albrivas.broadcastbottom.ui.settings
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.albrivas.broadcastbottom.databinding.FragmentSettingsBinding
+import org.koin.androidx.scope.lifecycleScope
+import org.koin.androidx.viewmodel.scope.viewModel
+
+class SettingsFragment : Fragment() {
+
+    private val viewModel: SettingsViewModel by lifecycleScope.viewModel(this)
+    private lateinit var binding: FragmentSettingsBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentSettingsBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+}
