+++ app/src/main/java/com/albrivas/broadcastbottom/ui/settings/SettingsViewModel.kt
@@ -1,0 +1,7 @@
+package com.albrivas.broadcastbottom.ui.settings
+
+import androidx.lifecycle.ViewModel
+
+class SettingsViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
