--- app/src/main/res/layout/alert_dialog_profile_information.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.appcompat.widget.LinearLayoutCompat xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="16dp">
-
-    <androidx.appcompat.widget.AppCompatEditText
-        android:id="@+id/editText"
-        android:layout_width="match_parent"
-        android:textSize="16sp"
-        android:backgroundTint="@android:color/transparent"
-        android:layout_height="wrap_content"
-        android:fontFamily="@font/ubuntu_regular" />
-
-</androidx.appcompat.widget.LinearLayoutCompat>
