+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -12,8 +12,8 @@
         android:title="@string/title_notifications" />
 
     <item
-        android:id="@+id/navigation_profile"
-        android:icon="@drawable/ic_person"
+        android:id="@+id/navigation_settings"
+        android:icon="@drawable/ic_settings"
         android:title="@string/title_settings" />
