+++ app/src/main/res/values/strings.xml
@@ -43,18 +43,5 @@
     <string name="facebook_app_id">3268637276559185</string>
     <string name="fb_login_protocol_scheme">fb3268637276559185</string>
 
-    <!-- Profile -->
-    <string name="profile_name">Nombre</string>
-    <string name="hint_name">Introduce tu nombre</string>
-    <string name="profile_email">Email</string>
-    <string name="hint_email">Introduce tu email</string>
-    <string name="profile_birthday">Fecha de nacimiento</string>
-    <string name="profile_phone">Teléfono</string>
-    <string name="hint_phone">Introduce tu teléfono</string>
-    <string name="date_format">%1$s-%2$s-%3$s</string>
-
-    <!-- Alert dialog -->
-    <string name="button_cancel">Cancelar</string>
-    <string name="button_edit">Editar</string>
 
 </resources>
